recon.py
A short python script developed for automating the subdomain scan, takeovers, generating screenshots and making a HTML report at the end for viewing the screenshots using pre made tools like aquatone,sublister and webscreenshot
Currently it is in development and for using it you have to play with the premade tools paths in the script. I will make it easy later on.